pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-llm-client"
version = "0.1.0"
description = "A unified async client for interacting with multiple LLM providers"
readme = "README.md"
authors = [
    {name = "skitsanos", email = "info@skitsanos.com"}
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["llm", "openai", "anthropic", "gpt", "claude", "ai", "machine learning", "ollama"]
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.22.0",
    "openai>=1.28.0",
    "python-dotenv>=1.0.0"
]

[project.urls]
"Homepage" = "https://github.com/skitsanos/unified-llm-client"
"Bug Tracker" = "https://github.com/skitsanos/unified-llm-client/issues"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["llm*"]
exclude = ["tests*", "examples*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"


# Development tool configurations
# @author: skitsanos
[tool.black]
line-length = 120
target-version = ["py310", "py311"]
include = '\\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
